requirements.txt
fastapi[standard]==0.115.4
uvicorn[standard]==0.32.0
aiosqlite==0.20.0
gunicorn==23.0.0
jinja2==3.1.6
python-multipart==0.0.7
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
alembic==1.14.0
reportlab==4.2.5
PyYAML==6.0.2
qrcode[pil]==7.4.2
Pillow==10.4.0
dnspython==2.6.1
itsdangerous==2.2.0
httpx==0.27.2
fastapi-users[sqlalchemy]==12.1.3
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
email-validator==2.1.1
setuptools>=70.0.0,<81
pytest==8.3.3
pytest-cov==5.0.0
pytest-xdist==3.5.0
pytest-asyncio==0.23.0
pytest-mock==3.12.0
coverage[toml]==7.4.0
black==24.10.0
pre-commit==3.7.1
bandit==1.7.10
libsass==0.23.0
mypy==1.11.2
isort==5.13.2
types-PyYAML==6.0.12.20240917
slowapi==0.1.9
asgi-correlation-id==4.3.1
python-json-logger==2.0.7
prometheus-client==0.20.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
defusedxml==0.7.1
pylint==3.2.6
ruff==0.14.4
autoflake==2.3.1
bokeh==3.5.0
lxml==5.3.0
python-dateutil==2.9.0
aiofiles==23.2.1
markdown==3.7
pygments==2.18.0
python-frontmatter==1.1.0
python-slugify==8.0.4
